feat: migrate Convert to Word to turn actions and make buttons permanent
- Moved "Convert to Word" from individual AI message bubbles to turn-level actions.
- Ensured Word export uses the same aggregated content as the Copy action.
- Removed hover-trigger logic to keep action buttons permanently visible.
- Removed redundant AI message copy buttons in MessagetText.tsx.
